番組情報表示関係のオプションを追加
・チューナー画面などでツールチップに表示出来るようにした
・チューナー画面などに「番組詳細タブを選択した状態で予約ダイアログを開く」を追加

チューナー画面で番組情報見れたら便利かと思ったが、いまいち。

+++ EpgTimer/EpgTimer/Menu/CmdExeReserve.cs
@@ -17,7 +17,7 @@ namespace EpgTimer
         public void SetFuncGetEpgEventList(Func<List<EpgEventInfo>> f) { _getEpgEventList = f; }
         public void SetFuncSelectSingleSearchData(Func<bool, SearchItem> f) { _selectSingleSearchData = f; }
 
-        public byte EpgInfoOpenMode { get; set; }
+        public Int32 EpgInfoOpenMode { get; set; }
         public RecSettingView recSettingView { get; set; }
 
         protected override int ItemCount
